+++ CHANGELOG.rst
@@ -26,6 +26,7 @@ Changelog
    - :func:`ndonnx.triu`
    - :func:`ndonnx.linspace`
 - Custom dtypes can now provide implementations for :func:`ndonnx.where`.
+- The :class:`ndonnx._experimental.UniformShapeOperations` now provides implementations of shape operations that are generic across all data types where each constituent field has the same shape (that of the overall array).
 
 **Other changes**
 
@@ -36,9 +37,6 @@ Changelog
 
 - Numerical operations like :func:`sin` now raise :class:`UnsupportedOperationError` when invoked using invalid data types like ``ndx.utf8`` rather than implicitly casting.
 - Fixes bug causing a promotion error when implementing numerical operations like :func:`add` that involve type promotion.
-
-**Bug fixes**
-
 - Fixes scalar promotion logic to more accurately reflect the Array API standard. Promotion requires at least one array to be present and scalars adopt the dtype of the arrays being promoted with it. `ndx.utf8` and `ndx.nutf8` cannot be promoted with any other dtypes.